replaced if tree with visitor in resolveAllStaticInitializerCallsFromSourceMethod
